requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fontawesomefree==6.5.1
PyJWT==2.10.1
sqlparse==0.5.2
tzdata==2024.2
